idea_forge/requirements.txt
crewai==1.9.3
pydantic==2.11.9
langchain==0.3.27
langchain-google-genai==2.1.10
google-genai==1.40.0
duckduckgo-search==8.1.1
beautifulsoup4==4.13.5
python-dotenv==1.1.1
